pyproject.toml
[tool.black]
line-length = 88
target-version = ['py310', 'py311']

[tool.coverage.run]
omit = ["*/tests/*", "members/migrations"]

[tool.poetry]
name = "forenings_medlemmer"
version = "1.2"
description = "Forenings system for Coding Pirates"
authors = ["Kristoffer Rath Hansen <kristoffer@codingpirates.dk>"]

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.3"
dj-database-url = "^1.3"
dj-email-url = "1.0.6"
django = "^4.2"
django-bootstrap5 = "^21.3"
django-cors-headers = "3.5.0"
django-crispy-forms = "1.9.2"
django-cron = "0.6.0"
django-extensions = "3.2.1"
environs = "8.0.0"
factory-boy = "3.0.1"
flake8 = "3.8.4"
freezegun = "1.2.2"
graphene_django = "3.0.0"
gunicorn = "20.1.0"
psycopg2 = "2.9.5"
quickpay-api-client = "2.0.1"
sentry-sdk = "1.14"
tqdm = "4.53.0"
whitenoise = "5.2.0"
pydotplus = "2.0.2"
selenium = "4.9.0"
crispy-bootstrap5 = "^0.4"
pytz = "^2022.7"
